+++ packages/panel-server/src/routes/management.js
@@ -7,7 +7,6 @@ import certsRoutes from './management/certs.js';
 import tunnelRoutes from './management/tunnels.js';
 import sitesRoutes from './management/sites.js';
 import invitationRoutes from './management/invitations.js';
-import shellRoutes from './management/shell.js';
 import pluginRoutes from './management/plugins.js';
 
 export default async function managementRoutes(fastify, _opts) {
@@ -21,6 +20,5 @@ export default async function managementRoutes(fastify, _opts) {
   await fastify.register(usersRoutes);
   await fastify.register(certsRoutes);
   await fastify.register(invitationRoutes);
-  await fastify.register(shellRoutes);
   await fastify.register(pluginRoutes);
 }